+++ padeopsIO/wake_utils.py
@@ -511,62 +511,3 @@ def compute_vort_xrt(sl_dict):
     })
 
 
-# ------------------- IO ----------------
-
-
-def read_list(dir_name): 
-    """
-    Outputs a list of BudgetIO objects from a directory name by reading the associated
-    file `./Runs.csv`
-    
-    Arguments
-    ---------
-    dir_name (str) : directory path containing a file Runs.csv
-    
-    Returns
-    -------
-    run_list : list of BudgetIO objects
-    """
-    
-    # read runs
-    # https://stackoverflow.com/questions/24662571/python-import-csv-to-list
-    with open(os.path.join(dir_name, 'Runs.csv')) as f:
-        reader = csv.reader(f)
-        runs = list(reader)
-
-    run_list = []
-
-    for run in runs: 
-        run_list.append(padeopsIO.BudgetIO(run[0], padeops=True, verbose=False))
-        
-    return run_list
-
-
-def key_search_r(nested_dict, key):
-    """
-    Copied from budgetkey.py
-    
-    Performs a recursive search for the dictionary key `key` in any of the dictionaries contained 
-    inside `nested_dict`. Returns the value of nested_dict[key] at the first match. 
-    
-    Parameters
-    ----------
-    nested_dict (dict-like) : dictionary [possibly of dictionaries]
-    key (str) : dictionary key to match
-    
-    Returns
-    -------
-    nested_dict[key] if successful, None otherwise. 
-    """
-    
-    try: 
-        for k in nested_dict.keys(): 
-            if k == key: 
-                return nested_dict[k]
-            else: 
-                res = key_search_r(nested_dict[k], key)
-                if res is not None: 
-                    return res
-        
-    except AttributeError as e: 
-        return
